requirements.txt
pydub>=0.25.1
google-genai>=1.0.0
rapidfuzz>=3.5.0
python-dotenv>=1.0.0

